requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
networkx==3.2.1
matplotlib==3.8.2
PyQt6==6.6.1
pandas==2.2.0 